pytest: Separate task collection tests
